extends chat service class over 2 files
