+++ backend/src/chat/chat.service.ts
@@ -1,101 +1,21 @@
 import { Inject, Injectable } from '@nestjs/common';
 import { UsersService } from 'src/users/users.service';
-import { IMessage, IUser } from './properties';
+import { IMessage } from './properties';
 import { ChatDAO } from './chat.dao';
-import { Socket, Server } from 'socket.io';
+import { Server } from 'socket.io';
 import { IGameUser } from 'src/games/properties';
 import { GamesService } from 'src/games/games.service';
+import { ChatServiceBase } from './chat.servicebase';
 
 @Injectable()
-export class ChatService {
+export class ChatService extends ChatServiceBase {
   constructor(
     @Inject(UsersService)
-    private userService: UsersService,
+    protected userService: UsersService,
     @Inject(ChatDAO)
-    private chatDao: ChatDAO,
+    protected chatDao: ChatDAO,
   ) {
-    this.initializeMap();
-  }
-
-  activeClients: Map<string, IGameUser[]> = new Map<string, IGameUser[]>();
-  opponents: Map<string, string> = new Map<string, string>();
-
-  async initializeMap(): Promise<void> {
-    const list = await this.chatDao.getAllChannels();
-    list.forEach((item) => {
-      this.activeClients.set(item, []);
-    });
-  }
-
-  private updateActiveClients(data: IMessage, client: Socket) {
-    for (const [key, value] of this.activeClients) {
-      const tmp = value.filter((c) => c.user.name !== data.user.name);
-      this.activeClients.set(key, tmp);
-    }
-    if (!this.activeClients.has(data.room))
-      this.activeClients.set(data.room, []);
-    this.activeClients.get(data.room).push({ user: data.user, socket: client });
-  }
-
-  private getUser(name: string, room: string): IGameUser {
-    return this.activeClients.get(room).find((user) => user.user.name === name);
-  }
-
-  async getChats(userId: string): Promise<string[]> {
-    let res: string[] = [];
-    try {
-      const user = await this.userService.findOneByName(userId);
-      res = await this.chatDao.getRawUserChannels(user.id);
-    } catch (error) {
-      console.log(error);
-    }
-    return res;
-  }
-
-  async addChat(
-    userId: string,
-    chatName: string,
-    client: Socket,
-  ): Promise<string[]> {
-    const res: string[] = [];
-    try {
-      const user = await this.userService.findOneByName(userId);
-      await this.chatDao.saveChannel(chatName, userId);
-      client.join(chatName);
-      return await this.chatDao.getRawUserChannels(user.id);
-    } catch (error) {
-      console.log(error);
-    }
-    return res;
-  }
-
-  async removeChat(userId: string, chat: string): Promise<void> {
-    try {
-      const user = await this.userService.findOneByName(userId);
-      await this.chatDao.removeUserFromChannel(chat, user);
-    } catch (error) {
-      console.log(error);
-    }
-  }
-
-  async joinRoom(data: IMessage, client: Socket): Promise<string[]> {
-    let res: string[] = [];
-    try {
-      const channel = await this.chatDao.getChannelByTitle(data.room);
-      client.join(data.room);
-      this.updateActiveClients(data, client);
-      await this.chatDao.addUserToChannel(data.room, data.user.name);
-      await this.chatDao.saveMessageToChannel({
-        user: data.user,
-        input: 'joined room',
-        room: data.room,
-      });
-      client.to(data.room).emit('message', 'oheinzel: joined room');
-      res = await this.chatDao.getRawChannelMessages(channel.id);
-    } catch (error) {
-      console.log(error);
-    }
-    return res;
+    super(userService, chatDao);
   }
 
   async handleMessage(data: IMessage, server: Server, gameServ: GamesService) {
@@ -127,7 +47,7 @@ export class ChatService {
     server: Server,
     gameServ: GamesService,
   ) {
-    if (this.getUser(data.user.name, data.room)) {
+    if (this.opponents.get(data.user.name)) {
       await this.gameAccept(data, server, gameServ);
       return;
     }
